+++ lnmpkit/cmd/lnmpkit/helper.go
@@ -1,5 +1,0 @@
-package main
-
-func helper()  {
-
-}