+++ lnmpkit/server/mysql/helper.go
@@ -1,7 +1,0 @@
-package mysql
-
-import "fmt"
-
-func helper()  {
-	fmt.Println("server mysql help info")
-}