+++ windows/lnmp-laravel.ps1
@@ -10,14 +10,16 @@ if ($args -contains 'new' ){
   }
 }
 
+$global:LARAVEL_PATH=$args[1]
+
 if (!(Test-Path ${LARAVEL_PATH})){
   echo ""
-  echo "${LARAVEL_PATH} not existing"
+  echo "${LARAVEL_PATH} not existing, laravel is install..."
   echo ""
 # docker run --init -it --rm `
 #     --mount type=bind,src=$PWD,target=/app `
 #     --mount src=lnmp_composer_cache-data,target=/tmp/cache `
-#     $LNMP_PHP_IMAGE `
+#     khs1994/php:7.2.12-composer-alpine `
 #     laravel $args
 
 composer create-project --prefer-dist laravel/laravel=5.6.* "$LARAVEL_PATH"
@@ -29,8 +31,6 @@ composer create-project --prefer-dist laravel/laravel=5.6.* "$LARAVEL_PATH"
 }
 
 if ($args -contains 'new' ){
-  $global:LARAVEL_PATH=$args[1]
-
   cd ${LARAVEL_PATH}
 
   . "$PSScriptRoot/lnmp-laravel-init.ps1"