+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 
 ### Changes:
 
+* Remove composer from `khs1994/php:x.y.z-fpm-alpine`, you can access composer with `khs1994/php:x.y.z-composer-alpine`.
+* Move `docker-cluster.*.yml` to cluster folder.
+* Windows Container based `mcr.microsoft.com/windows/servercore:1809`
+
 ### Updates:
 
 * [Redis `5.0.1`](https://raw.githubusercontent.com/antirez/redis/5.0/00-RELEASENOTES)